% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronograph-package.R
\docType{package}
\name{chronograph-package}
\alias{chronograph}
\alias{chronograph-package}
\title{chronograph: Chronological Networks}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeroe/chronograph}
  \item Report bugs at \url{https://github.com/joeroe/chronograph/issues}
}

}
\keyword{internal}
