% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronologr-package.R
\docType{package}
\name{chronologr-package}
\alias{chronologr}
\alias{chronologr-package}
\title{chronologr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeroe/chronologr}
  \item Report bugs at \url{https://github.com/joeroe/chronologr/issues}
}

}
\keyword{internal}
