% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crg_add_synchronism.R
\name{crg_add_synchronism}
\alias{crg_add_synchronism}
\alias{crg_equal_to}
\title{Temporal synchronisms}
\usage{
crg_add_synchronism(chronograph, relation, x, y)

crg_equal_to(chronograph, x, y)
}
\arguments{
\item{chronograph}{A \code{\link[=chronograph]{chronograph()}} model}

\item{relation}{Character. One of the types of relations defined below.}

\item{x, y}{Numeric indices or character labels of the two periods referred to
by \code{relation}}
}
\description{
Functions for specifying known temporal relations between two periods within
a chronological graph model.
}
\examples{
model <- chronograph(periods = c("X", "Y"))
crg_add_synchronism(chronograph, "equal_to", "X", "Y")
crg_equal_to(chronograph, "X", "Y")
}
