% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crg_period.R
\name{crg_period}
\alias{crg_period}
\title{Chronological periods}
\usage{
crg_period(
  start_lower = yr(),
  start_upper = yr(),
  end_lower = yr(),
  end_upper = yr(),
  duration_lower = integer(),
  duration_upper = integer()
)
}
\arguments{
\item{start_lower}{\code{\link[=yr]{yr()}} or integer vector (see details) with \emph{terminus post quem} dates of the start of the period.}

\item{start_upper}{\code{\link[=yr]{yr()}} or integer vector (see details) with \emph{terminus ante quem} dates of the start of the period.}

\item{end_lower}{\code{\link[=yr]{yr()}} or integer vector (see details) with \emph{terminus post quem} dates of the end of the period.}

\item{end_upper}{\code{\link[=yr]{yr()}} or integer vector (see details) with \emph{terminus ante quem} dates of the end of the period.}

\item{duration_lower}{Integer vector of lower bounds of the duration of the period.}

\item{duration_upper}{Integer vector of upper bounds of the duration of the period.}
}
\value{
A vector of periods with S3 class \code{crg_period}.
}
\description{
This S3 vector class describes periods of time through contraints on their
start, end and duration. Uncertainty about each of these elements is
expressed as a lower or upper bound, either of both or which can be \code{NA}.
}
\details{
The duration of a period can be zero, representing a singular event.
}
\examples{
crg_period()
}
