% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crg_relation.R
\name{crg_relation}
\alias{crg_relation}
\alias{crg_equals}
\title{Temporal relations}
\usage{
crg_relation(type = "equals", x, y)

crg_equals(x, y)
}
\arguments{
\item{type}{Character. One of the types of relations defined below.}

\item{x}{Unique character label of the first period}

\item{y}{Unique character label of the second period}
}
\description{
Functions expressing different types of temporal relation between two
periods.
}
\examples{
# Expresses that the periods labelled 'X' and 'Y' are equal:
crg_relation("equals", "X", "Y")

# Which is equivalent to:
crg_relation("equals", "X", "Y")
}
