% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crg_synchronism.R
\name{synchronism}
\alias{synchronism}
\title{Temporal synchronism}
\usage{
synchronism(relation = crg_equals, x = character(), y = character())
}
\arguments{
\item{relation}{Function describing the type of relation, see \code{\link[=crg_relation]{crg_relation()}}}

\item{x}{Unique character label of the first period}

\item{y}{Unique character label of the second period}
}
\value{
A vector of temporal synchronisms
}
\description{
A synchronism defines a known relation between two periods.
}
\details{
\code{x} and \code{y} are \link[vctrs:theory-faq-recycling]{recycled to a common length}.
}
\examples{
synchronism(crg_equals, "A", "B")
}
